+++ routes/web.php
@@ -92,4 +92,8 @@ Route::post('/telescope-api/toggle-recording', 'RecordingController@toggle');
 // Clear Entries...
 Route::delete('/telescope-api/entries', 'EntriesController@destroy');
 
+// Prune Entries...
+Route::get('/telescope-api/entries/prune', 'EntriesController@pruneStatus');
+Route::post('/telescope-api/entries/prune', 'EntriesController@prune');
+
 Route::get('/{view?}', 'HomeController@index')->where('view', '(.*)')->name('telescope');